adjusted builtins functions. added bultins files to makefile and cleaned up builtins.h

+++ sources/builtins/unset.c
@@ -11,7 +11,7 @@ void unset_env_var(char *key, t_env **env_list)
 	key_len = ft_strlen(key);
 	while (temp != NULL)
 	{
-		if (ft_strlen(temp->key) == key_len && ft_strncmp(temp->key, key, key_len) == 0)
+		if ((int)ft_strlen(temp->key) == key_len && ft_strncmp(temp->key, key, key_len) == 0)
 		{
 			if (prev == NULL)
 				*env_list = temp->next;
